api: rename the last file
